Reorganize controller directories
