+++ pkg/controllers/controllers.go
@@ -22,18 +22,18 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/aws/karpenter-core/pkg/cloudprovider"
-	"github.com/aws/karpenter-core/pkg/controllers/consistency"
-	"github.com/aws/karpenter-core/pkg/controllers/counter"
 	"github.com/aws/karpenter-core/pkg/controllers/deprovisioning"
 	"github.com/aws/karpenter-core/pkg/controllers/leasegarbagecollection"
+	"github.com/aws/karpenter-core/pkg/controllers/machine/consistency"
 	nodeclaimdisruption "github.com/aws/karpenter-core/pkg/controllers/machine/disruption"
 	nodeclaimgarbagecollection "github.com/aws/karpenter-core/pkg/controllers/machine/garbagecollection"
 	nodeclaimlifecycle "github.com/aws/karpenter-core/pkg/controllers/machine/lifecycle"
 	nodeclaimtermination "github.com/aws/karpenter-core/pkg/controllers/machine/termination"
+	metricsnode "github.com/aws/karpenter-core/pkg/controllers/metrics/node"
 	metricspod "github.com/aws/karpenter-core/pkg/controllers/metrics/pod"
 	metricsprovisioner "github.com/aws/karpenter-core/pkg/controllers/metrics/provisioner"
-	metricsstate "github.com/aws/karpenter-core/pkg/controllers/metrics/state"
-	"github.com/aws/karpenter-core/pkg/controllers/provisioner"
+	"github.com/aws/karpenter-core/pkg/controllers/provisioner/counter"
+	"github.com/aws/karpenter-core/pkg/controllers/provisioner/hash"
 	"github.com/aws/karpenter-core/pkg/controllers/provisioning"
 	"github.com/aws/karpenter-core/pkg/controllers/state"
 	"github.com/aws/karpenter-core/pkg/controllers/state/informer"
@@ -58,10 +58,9 @@ func NewControllers(
 
 	return []controller.Controller{
 		p,
-		metricsstate.NewController(cluster),
 		deprovisioning.NewController(clock, kubeClient, p, cloudProvider, recorder, cluster),
 		provisioning.NewController(kubeClient, p, recorder),
-		provisioner.NewProvisionerController(kubeClient),
+		hash.NewProvisionerController(kubeClient),
 		informer.NewDaemonSetController(kubeClient, cluster),
 		informer.NewNodeController(kubeClient, cluster),
 		informer.NewPodController(kubeClient, cluster),
@@ -70,6 +69,7 @@ func NewControllers(
 		termination.NewController(kubeClient, cloudProvider, terminator, recorder),
 		metricspod.NewController(kubeClient),
 		metricsprovisioner.NewController(kubeClient),
+		metricsnode.NewController(cluster),
 		counter.NewProvisionerController(kubeClient, cluster),
 		consistency.NewMachineController(clock, kubeClient, recorder, cloudProvider),
 		nodeclaimlifecycle.NewMachineController(clock, kubeClient, cloudProvider, recorder),
