+++ pkg/controllers/machine/consistency/nodeshape.go
